% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_matrix.R
\name{add_triangle}
\alias{add_triangle}
\title{Determines whether or not a value is in the upper or lower triangle
of a correlation matrix.}
\usage{
add_triangle(df)
}
\arguments{
\item{df}{a data frame or tibble.}
}
\value{
a data frame with labeled upper and lower triangles
}
\description{
This is a helper function for long_cors
}
