% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_matrix.R
\name{get_cor_df}
\alias{get_cor_df}
\title{Get a data frame of correlations and p values}
\usage{
get_cor_df(df)
}
\arguments{
\item{df}{a data frame or tibble.}
}
\value{
correlations and p values
}
\description{
This function is indented to store correlations and p values in an easy-to-
work-with dataframe.
}
