% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_effects.R
\name{get_model_predictions}
\alias{get_model_predictions}
\title{Return a dataframe of model predictions}
\usage{
get_model_predictions(model, data = NULL, label = NULL, type = "response")
}
\value{
A dataframe of model predictions.
}
\description{
This is an internal function used by other graphing functions.
}
