% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_effects.R
\name{get_predictions}
\alias{get_predictions}
\title{Get predictions from a model}
\usage{
get_predictions(m, varying, full_range_var = NULL, sd_to_graph = 1)
}
\arguments{
\item{m}{A linear model. Standard OLS or logistic regression will work.}

\item{varying}{A vector of the names of the terms in the model to vary.}

\item{full_range_var}{Should one of the variables be graphed continuously on the x axis? If so,
you can write its name here.}

\item{sd_to_graph}{When making predictions for continuous variables, the function will make them
at +/- this many standard deviations.}
}
\value{
A data frame of model predictions.
}
\description{
This function returns a dataframe of model predictions, varying the variables you specify
and holding the others constant at zero. Accordingly, it's important to center continuous
variables and set up appropriate contrasts for factor variables before running this
function.
}
