% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_effects.R
\name{graph_all_effects}
\alias{graph_all_effects}
\title{Graph all of the effects in a model that meet a significance threshold.}
\usage{
graph_all_effects(
  m,
  p_threshold = 0.05,
  x_pref = NULL,
  fill_pref = NULL,
  facet_pref = NULL,
  continuous_x = FALSE,
  colors = NULL
)
}
\arguments{
\item{m}{A linear model. Standard OLS or logistic regression will work.}

\item{p_threshold}{All p values less than or equal to this threshold will be graphed.}

\item{x_pref}{An optional vector of preferences for the x axis.}

\item{fill_pref}{An optional vector of preferences for the fill of the graph.}

\item{facet_pref}{An optional vector of preferences for the facets of the graph.}

\item{continuous_x}{If true, a line graph will be created for any continuous variables, rather than bar graphs with predicted values at +1/-1 SD.}

\item{colors}{An optional vector of colors for the bars and lines of graphs.}
}
\value{
A set of ggplot graphs.
}
\description{
This function is intended to quickly view and model output (e.g. interactions).
All factor variables should be given contrasts.
}
