% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_matrix.R
\name{graph_cors}
\alias{graph_cors}
\title{Graph a data frame of correlations and p values, in long format.}
\usage{
graph_cors(df)
}
\arguments{
\item{df}{a data frame or tibble.}
}
\value{
a ggplot of correlations
}
\description{
This is designed to work with long_cors in the pipe.
}
