% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_effects.R
\name{graph_effect}
\alias{graph_effect}
\title{Graph a single effect, given a model.}
\usage{
graph_effect(
  model,
  x,
  fill = NULL,
  facet = NULL,
  facet2 = NULL,
  continuous_x = FALSE,
  colors = NULL
)
}
\arguments{
\item{x}{the name of the variable that should be graphed on the x axis.}

\item{fill}{the (optional) name of the variable that should be graphed as a fill.}

\item{facet}{the (optional) name of the variable that should be graphed as a facet.}

\item{facet2}{the (optional) name of the variable that should be graphed as a second facet.}

\item{continuous_x}{If true, a line graph will be created for any continuous variables, rather than bar graphs with predicted values at +1/-1 SD.}

\item{colors}{An optional vector of colors for the bars and lines of graphs.}

\item{m}{A linear model. Standard OLS or logistic regression will work.}
}
\value{
A ggplot line or bar graph.
}
\description{
This function is intended to quickly graph a single set of predictions, given a model.
}
