% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphEffect.R
\name{graph_interaction}
\alias{graph_interaction}
\title{Generate a bar graph of a specified interaction}
\usage{
graph_interaction(d, m, x, fill, xlevels = c(-1, 1), fill2, x2)
}
\arguments{
\item{d}{The dataframe from which the model was generated}

\item{m}{A model}

\item{x}{The variable that you want on the x axis}

\item{fill}{The variable that you want as a moderator}

\item{xlevels}{The levels at which you want to vary any continuous parameters}

\item{fill2}{An optional second moderator, graphed as a fill}

\item{x2}{An optional second independent variable, graphed on the x axis}
}
\value{
A ggplot of of model predictions for the interaction
}
\description{
This function generates a bar graph of an interaction, given a model,
one or two x variables, and one or two moderators.
}
