% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_effects.R
\name{graph_predictions}
\alias{graph_predictions}
\title{Graph a dataframe of model predictions.}
\usage{
graph_predictions(
  predictions,
  geom = "bar",
  x = NULL,
  fill = NULL,
  facet = NULL,
  facet2 = NULL,
  colors = NULL
)
}
\arguments{
\item{predictions}{A dataframe of model predictions. The y variable must be in the first column.}

\item{x}{The variable to graph on the x axis.}

\item{fill}{The (optional) variable to graph as a fill.}

\item{facet}{The (optional) variable to graph as a facet}

\item{facet2}{The (optional) variable to graph as a second facet (for 4-way interactions)}

\item{color}{An optional vector of colors for bars or lines.}
}
\value{
A ggplot (bar or line graph).
}
\description{
This function returns a ggplot
}
