% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_fiml.R
\name{lm_fiml}
\alias{lm_fiml}
\title{Run a model that has lm syntax in lavaan with FIML}
\usage{
lm_fiml(model)
}
\arguments{
\item{model}{An lm model}
}
\value{
Model output
}
\description{
This function is intended to allow you to avoid manually creating interaction
terms when you want to handle missing data with FIML.
}
