% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_matrix.R
\name{long_cors}
\alias{long_cors}
\title{Get a data frame of correlations, in long format.}
\usage{
long_cors(df)
}
\arguments{
\item{df}{a data frame or tibble.}
}
\value{
correlations and p values
}
\description{
A handy function to get lots of correlations in a tidy format
}
