% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphEffect.R
\name{modelEffect}
\alias{modelEffect}
\title{Generate a table of estimates given a model}
\usage{
modelEffect(m, d, varying, xlevels = c(-1, 1))
}
\arguments{
\item{m}{A model}

\item{d}{The dataframe from which the model was generated}

\item{varying}{A vector of the model parameters that you wish to vary}

\item{xlevels}{The levels at which you want to vary any continuous parameters}

\item{...}{up to 8 linear models, produced by lm}
}
\value{
A dataframe of model predictions, standard errors, and confidence intervals
}
\description{
This function generates estimates given a model, varying specified parameters.
}
