% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qualtrics.R
\name{read_qualtrics}
\alias{read_qualtrics}
\title{This function reads in data given a csv in the format that qualtrics generates,
with the header row, a row of information about the variables, and then the data.
Make sure you export the data from qualtrics removing line breaks, or this function
won't work properly.}
\usage{
read_qualtrics(filePath)
}
\value{
a data frame.
}
\description{
This function reads in data given a csv in the format that qualtrics generates,
with the header row, a row of information about the variables, and then the data.
Make sure you export the data from qualtrics removing line breaks, or this function
won't work properly.
}
