% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_effects.R
\name{set_x_fill_facet}
\alias{set_x_fill_facet}
\title{Graph a dataframe of model predictions.}
\usage{
set_x_fill_facet(predictions, x_pref, fill_pref, facet_pref)
}
\arguments{
\item{predictions}{A dataframe of model predictions. The y variable must be in the first column.}

\item{x_pref}{An optional vector of preferences for the x axis.}

\item{fill_pref}{An optional vector of preferences for the fill of the graph.}

\item{facet_pref}{An optional vector of preferences for the facets of the graph.}
}
\value{
A list with x, fill, and facet decisions.
}
\description{
This function determines the proper x, fill, facet, and facet2, given a set of
preferences. It is used by other graphing functions.
}
