% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setwd_toSource.R
\name{setwd_toSource}
\alias{setwd_toSource}
\title{This fucntion sets the working directory to the source file location, using
the RStudio API.}
\usage{
setwd_toSource()
}
\value{
The working directory is printed to the console.
}
\description{
This fucntion sets the working directory to the source file location, using
the RStudio API.
}
