% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepTable.R
\name{stepTable}
\alias{stepTable}
\title{Make a Stepwise Regression Table}
\usage{
stepTable(..., Betas = TRUE, ModelStats = FALSE)
}
\arguments{
\item{...}{up to 8 linear models, produced by lm}

\item{Betas}{whether standardized betas should be printed}

\item{ModelStats}{whether r squared, etc. should be printed}
}
\value{
The regression table is printed to the viewer using the "Kable" function from knitr
}
\description{
This function creates a summary table for a linear model, and outputs it to the
viewer in HTML Format to facilitate copyin and pasting. By default it outputs
standardized betas for regression coefficients, and it can accept
up to 8 models as arguments, which makes it ideal for displaying stepwise models.
}
