% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphEffect.R
\name{graph_effect}
\alias{graph_effect}
\title{Generate a bar graph, given a set of model predictions}
\usage{
graph_effect(d, x, y, fill, fill2, x2)
}
\arguments{
\item{d}{A dataframe of model predictions, e.g. generated by the modelEffect function}

\item{x}{The variable that you want on the x axis}

\item{y}{The variable that you want on the y axis}

\item{fill}{The variable that you want as a moderator}

\item{fill2}{An optional second moderator, graphed as a fill}

\item{x2}{An optional second independent variable, graphed on the x axis}
}
\value{
A ggplot of of model predictions for the interaction
}
\description{
This function generates a bar graph of an interaction, given a dataframe of model
predictions.
}
