% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepTable.R
\name{stepTable}
\alias{stepTable}
\title{Make a Stepwise Regression Table}
\usage{
stepTable(..., Betas = TRUE, ModelStats = FALSE, Output = c("viewer",
  "markdown"))
}
\arguments{
\item{...}{up to 8 linear models, produced by lm}

\item{Betas}{whether standardized betas should be printed}

\item{ModelStats}{whether r squared, etc. should be printed}

\item{Output}{whether output should be displayed in the viewer, or formatted with "Kable" for markdown}
}
\value{
The regression table
}
\description{
This function creates a summary table for a linear model, and outputs it to the
viewer in HTML Format to facilitate copying and pasting by default. By default it outputs
standardized betas for regression coefficients, and it can accept
up to 8 models as arguments, which makes it ideal for displaying stepwise models.
}
